@@ -4,13 +4,14 @@
   "license": "MIT",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve --port 8000",
-    "build": "ng build",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e",
-    "build:universal": "ng build --prod && ng build --prod --app 1 --output-hashing=false",
-    "serve:universal": "npm run build:universal && node server"
+    "serve": "ng serve --port 8000",
+    "serve:dynamic": "node --experimental-modules dist/server.js",
+    "start": "npm run ssr",
+    "ssr": "npm run build:dynamic && npm run serve:dynamic",
+    "build:dynamic": "npm run build:client && npm run build:server && npm run webpack:server",
+    "build:client": "ng build --prod",
+    "build:server": "ng build --prod --app 1 --output-hashing=false",
+    "webpack:server": "webpack --config webpack.server.config.js -p --progress"
   },
   "private": true,
   "dependencies": {
@@ -26,9 +27,17 @@
     "@angular/router": "^5.0.3",
     "@nguniversal/express-engine": "^5.0.0-beta.5",
     "@nguniversal/module-map-ngfactory-loader": "^5.0.0-beta.5",
+    "apollo-angular": "1.0.0-beta.1",
+    "apollo-angular-link-http": "1.0.0-beta.1",
+    "apollo-cache-inmemory": "1.0.0",
+    "apollo-client": "2.0.1",
+    "apollo-link": "1.0.0",
     "core-js": "^2.4.1",
     "express": "^4.16.2",
-    "rxjs": "^5.5.2",
+    "graphql-tag": "^2.5.0",
+    "rxjs": "^5.4.2",
+    "ts-loader": "^2.3.7",
+    "webpack-node-externals": "^1.6.0",
     "zone.js": "^0.8.14"
   },
   "devDependencies": {
@@ -37,7 +46,8 @@
     "@angular/language-service": "^5.0.3",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
+    "@types/node": "~8.0.55",
+    "@types/zen-observable": "^0.5.3",
     "codelyzer": "~3.2.0",
     "jasmine-core": "~2.6.2",
     "jasmine-spec-reporter": "~4.1.0",
@@ -50,6 +60,7 @@
     "protractor": "~5.1.2",
     "ts-node": "~3.2.0",
     "tslint": "~5.7.0",
-    "typescript": "~2.4.2"
+    "graphql": "^0.11.7",
+    "typescript": "2.5.3"
   }
 }
