@@ -5,7 +5,25 @@ import { BehaviorSubject } from 'rxjs/BehaviorSubject';
 
 import { ProjectShort } from './project-short.model';
 import { environment } from '../../environments/environment';
+import gql from 'graphql-tag';
+import { Apollo } from 'apollo-angular';
 
+const gqltag = gql(`query {
+  sectionByDomain(domain: "works.novel.tl") {
+    sectionId
+    url
+    visibleProjects {
+      projectId
+      url
+      title
+      status
+      translationStatus
+      banner {
+        url
+      }
+    }
+  }
+}`);
 
 @Injectable()
 export class ProjectsService {
@@ -13,28 +31,33 @@ export class ProjectsService {
   private readonly RESULT_KEY = makeStateKey<ProjectShort[]>('projects_store');
 
 
-  private data:BehaviorSubject<ProjectShort[]> = new BehaviorSubject<ProjectShort[]>(this.getProjects());
+  private data: BehaviorSubject<ProjectShort[]>; // = new BehaviorSubject<ProjectShort[]>(this.getProjects());
 
   projects = this.data.asObservable();
 
-  constructor(private http: HttpClient, private state: TransferState,) { }
+  constructor(private http: HttpClient, private state: TransferState, private apollo: Apollo) { }
 
   private fetch() {
-    this.http
-      .get(`${environment.apiHost}/projects?showHidden=false`)
-      .subscribe((res:ProjectShort[]) => {
-        this.state.set<ProjectShort[]>(this.RESULT_KEY, res);
-        this.data.next(res);
-      });
+    // this.http
+    //   .get(`${environment.apiHost}/projects?showHidden=false`)
+    //   .subscribe((res: ProjectShort[]) => {
+    //     this.state.set<ProjectShort[]>(this.RESULT_KEY, res);
+    //     this.data.next(res);
+    //   });
+    this.apollo.query(gqltag).subscribe((res) => {
+      console.log(res);
+      // this.data = res;
+    });
   }
 
-  private getProjects():ProjectShort[] {
-    let data = this.state.get<ProjectShort[]>(this.RESULT_KEY, []);
+  private getProjects(): ProjectShort[] | null {
+    // let data = this.state.get<ProjectShort[]>(this.RESULT_KEY, []);
 
-    if (!this.state.hasKey(this.RESULT_KEY)) {
+    // if (!this.state.hasKey(this.RESULT_KEY)) {
       this.fetch();
-    }
+    // }
 
-    return data;
+    // return this.data;
+    return null;
   }
 }
